Centralize routing and dispatching functional test logic
