+++ test/Microsoft.AspNetCore.Mvc.FunctionalTests/Microsoft.AspNetCore.Mvc.FunctionalTests.csproj
@@ -31,7 +31,6 @@
     <ProjectReference Include="..\WebSites\BasicWebSite\BasicWebSite.csproj" />
     <ProjectReference Include="..\WebSites\ControllersFromServicesWebSite\ControllersFromServicesWebSite.csproj" />
     <ProjectReference Include="..\WebSites\CorsWebSite\CorsWebSite.csproj" />
-    <ProjectReference Include="..\WebSites\DispatchingWebSite\DispatchingWebSite.csproj" />
     <ProjectReference Include="..\WebSites\ErrorPageMiddlewareWebSite\ErrorPageMiddlewareWebSite.csproj" />
     <ProjectReference Include="..\WebSites\FilesWebSite\FilesWebSite.csproj" />
     <ProjectReference Include="..\WebSites\FormatterWebSite\FormatterWebSite.csproj" />
