--- test/WebSites/DispatchingWebSite/readme.md
@@ -1,4 +1,0 @@
-DispatchingWebSite
-===
-
-This web site illustrates how to use conventional and attribute dispatching.
