+++ test/WebSites/RoutingWebSite/StartupWithDispatching.cs
@@ -14,9 +14,9 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 
-namespace DispatchingWebSite
+namespace RoutingWebSite
 {
-    public class Startup
+    public class StartupWithDispatching
     {
         // Set up application services
         public void ConfigureServices(IServiceCollection services)
@@ -33,46 +33,26 @@ namespace DispatchingWebSite
         {
             app.UseDispatcher();
 
-            //app.UseMvc(routes =>
+            app.UseEndpoint();
+
+            //app.UseMvcWithEndpoint(routes =>
             //{
-            //    routes.MapAreaRoute(
+            //    routes.MapAreaEndpoint(
             //       "flightRoute",
             //       "adminRoute",
             //       "{area:exists}/{controller}/{action}",
             //       new { controller = "Home", action = "Index" },
             //       new { area = "Travel" });
 
-            //    routes.MapRoute(
+            //    routes.MapEndpoint(
             //        "ActionAsMethod",
             //        "{controller}/{action}",
             //        defaults: new { controller = "Home", action = "Index" });
 
-            //    routes.MapRoute(
+            //    routes.MapEndpoint(
             //        "RouteWithOptionalSegment",
             //        "{controller}/{action}/{path?}");
             //});
-
-            app.UseEndpoint();
         }
-
-        public static void Main(string[] args)
-        {
-            var host = CreateWebHostBuilder(args)
-                .ConfigureLogging((hostingContext, logging) =>
-                {
-                    logging.AddConsole();
-                })
-                .Build();
-
-            host.Run();
-        }
-
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            new WebHostBuilder()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseStartup<Startup>()
-                .UseKestrel()
-                .UseIISIntegration();
     }
 }
-
